refactor: change FileStorageOption's namespace
Move FileStorageOption to namespace 'CampusTrade.API.Options'
